Migrate Cache and Database types into backend-plugin-api
