+++ .changeset/nasty-pumas-kneel.md
@@ -1,0 +1,6 @@
+---
+'@backstage/backend-plugin-api': patch
+'@backstage/backend-common': patch
+---
+
+Migrated types related to `CacheService` and `DatabaseService` into backend-plugin-api.
