+++ packages/backend-app-api/api-report.md
@@ -4,14 +4,14 @@
 
 ```ts
 import { BackendFeature } from '@backstage/backend-plugin-api';
-import { CacheService } from '@backstage/backend-plugin-api';
 import { ConfigService } from '@backstage/backend-plugin-api';
-import { DatabaseService } from '@backstage/backend-plugin-api';
 import { ExtensionPoint } from '@backstage/backend-plugin-api';
 import { HttpRouterService } from '@backstage/backend-plugin-api';
 import { LifecycleService } from '@backstage/backend-plugin-api';
 import { LoggerService } from '@backstage/backend-plugin-api';
 import { PermissionsService } from '@backstage/backend-plugin-api';
+import { PluginCacheManager } from '@backstage/backend-common';
+import { PluginDatabaseManager } from '@backstage/backend-common';
 import { PluginEndpointDiscovery } from '@backstage/backend-common';
 import { RootLifecycleService } from '@backstage/backend-plugin-api';
 import { RootLoggerService } from '@backstage/backend-plugin-api';
@@ -34,7 +34,7 @@ export interface Backend {
 // @public (undocumented)
 export const cacheFactory: (
   options?: undefined,
-) => ServiceFactory<CacheService>;
+) => ServiceFactory<PluginCacheManager>;
 
 // @public (undocumented)
 export const configFactory: (
@@ -55,7 +55,7 @@ export interface CreateSpecializedBackendOptions {
 // @public (undocumented)
 export const databaseFactory: (
   options?: undefined,
-) => ServiceFactory<DatabaseService>;
+) => ServiceFactory<PluginDatabaseManager>;
 
 // @public (undocumented)
 export const discoveryFactory: (
